+++ config/test.exs
@@ -1,5 +1,3 @@
 use Mix.Config
 
-config :logger, backends: []
-
-config :coney, Coney.AMQPConnection, adapter: Coney.Test.FakeConnection, settings: %{}
+config :coney, adapter: Coney.FakeConnection, pool_size: 1, settings: %{}