+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [1.0.0]
+
+### Changes
+
+- `amqp` updated to version 1.0
+- `poison` removed from dependencies
+- Changed format of consumer `process/2` and `error_happened/3` functions
+- `error_happened/3` marked as optional
+- Added `Coney.Consumer` behaviour
+- `Coney.AMQPConnection` removed from configs
+- Logging removed
+- Added option `pool_size` - number of RabbitMQ connections.
+
 ## [0.4.3]
 
 ### Changes
@@ -19,17 +32,17 @@
 
 - `{:reject, reason}` return value:
   Reject message without redelivery.
-  
+
   ```elixir
   defmodule MyConsumer do
     def connection do
       #...
     end
-  
+
     def parse(payload) do
       String.to_integer(payload)
     end
-  
+
     def process(number) do
       if number <= 10 do
         {:ok, "Work done"}